# pkg/api/models/ok_body.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
)

// OkBody OkBody OkBody ok body
//
// swagger:model OkBody
type OkBody map[string]interface{}

// Validate validates this ok body
func (m OkBody) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this ok body based on context it is used
func (m OkBody) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}