# scripts/api/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Direktiv Open API Specification\nDirektiv Documentation can be found at https://docs.direktiv.io/",
    "title": "Direktiv API.",
    "contact": {
      "email": "info@direktiv.io"
    },
    "version": "1.0.0"
  },
  "host": "localhost",
  "paths": {
...
    "/api/logs": {
      "get": {
        "description": "Gets Direktiv Server Logs.\n",
        "tags": [
          "Logs"
        ],
        "summary": "Get Direktiv Server Logs",
        "operationId": "serverLogs",
        "responses": {
          "200": {
            "description": "successfully got server logs",
            "schema": {
              "$ref": "#/definitions/OkBody"
            }
          },
          "default": {
            "description": "an error has occurred",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/api/namespaces/{namespace}/instances/{instance}/logs": {
      "get": {
        "description": "Gets the logs of an executed instance.\n",
        "tags": [
          "Logs"
        ],
        "summary": "Gets Instance Logs",
        "operationId": "instanceLogs",
        "parameters": [
          {
            "type": "string",
            "description": "target namespace",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "target instance id",
            "name": "instance",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "namespace has been successfully created",
            "schema": {
              "$ref": "#/definitions/OkBody"
            }
          },
          "default": {
            "description": "an error has occurred",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/api/namespaces/{namespace}/instances/{instance}/vars": {
      "get": {
        "description": "Gets a list of variables in a instance.\n",
        "tags": [
          "Variables"
        ],
        "summary": "Get List of Instance Variable",
        "operationId": "getInstanceVariables",
        "parameters": [
          {
            "type": "string",
            "description": "target namespace",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "target instance",
            "name": "instance",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successfully got instance variables"
          }
        }
      }
    },
...
    "/api/namespaces/{namespace}/logs": {
      "get": {
        "description": "Gets Namespace Level Logs.\n",
        "tags": [
          "Logs"
        ],
        "summary": "Gets Namespace Level Logs",
        "operationId": "namespaceLogs",
        "parameters": [
          {
            "type": "string",
            "description": "target namespace",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successfully got namespace logs"
          }
        }
      }
    },
...
    "/api/namespaces/{namespace}/registries": {
      "get": {
        "description": "Gets the list of namespace registries.\n",
        "tags": [
          "Registries"
        ],
        "summary": "Get List of Namespace Registries",
        "operationId": "getRegistries",
        "parameters": [
          {
            "type": "string",
            "description": "target namespace",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successfully got namespace registries"
          }
        }
      },
      "post": {
        "description": "Delete a namespace container registry\n",
        "tags": [
          "Registries"
        ],
        "summary": "Delete a Namespace Container Registry",
        "operationId": "deleteRegistry",
        "parameters": [
          {
            "type": "string",
            "description": "target namespace",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "description": "Payload that contains registry data",
            "name": "Registry Payload",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "reg"
              ],
              "properties": {
                "reg": {
                  "description": "Target registry URL",
                  "type": "string"
                }
              },
              "example": {
                "data": "admin:8QwFLg%D$qg*",
                "reg": "https://prod.customreg.io"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfully delete namespace registry"
          }
        }
      }
    },
    "/api/namespaces/{namespace}/secrets": {
      "get": {
        "description": "Gets the list of namespace secrets.\n",
        "tags": [
          "Secrets"
        ],
        "summary": "Get List of Namespace Secrets",
        "operationId": "getSecrets",
        "parameters": [
          {
            "type": "string",
            "description": "target namespace",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successfully got namespace nodes",
            "schema": {
              "$ref": "#/definitions/OkBody"
            }
          },
          "default": {
            "description": "an error has occurred",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/api/namespaces/{namespace}/tree/{directory}?op=create-directory": {
      "put": {
        "description": "Creates a directory at the target path.\n",
        "tags": [
          "Directory"
        ],
        "summary": "Create a Directory",
        "operationId": "createDirectory",
        "parameters": [
          {
            "type": "string",
            "default": "create-directory",
            "description": "the operation for the api",
            "name": "op",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "target namespace",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "path to target directory",
            "name": "directory",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "directory has been created",
            "schema": {
              "$ref": "#/definitions/OkBody"
            }
          },
          "default": {
            "description": "an error has occurred",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/api/namespaces/{namespace}/tree/{nodePath}": {
      "get": {
        "description": "Gets Workflow and Directory Nodes at nodePath.\n",
        "tags": [
          "Node"
        ],
        "summary": "Get List of Namespace Nodes",
        "operationId": "getNodes",
        "parameters": [
          {
            "type": "string",
            "description": "target namespace",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "target path in tree",
            "name": "nodePath",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successfully got namespace nodes",
            "schema": {
              "$ref": "#/definitions/OkBody"
            }
          },
          "default": {
            "description": "an error has occurred",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/api/namespaces/{namespace}/tree/{node}?op=delete-node": {
      "delete": {
        "description": "Creates a directory at the target path.\n",
        "tags": [
          "Node"
        ],
        "summary": "Delete a node",
        "operationId": "deleteNode",
        "parameters": [
          {
            "type": "string",
            "default": "delete-node",
            "description": "the operation for the api",
            "name": "op",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "target namespace",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "path to target node",
            "name": "node",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "node has been deleted",
            "schema": {
              "$ref": "#/definitions/OkBody"
            }
          },
          "default": {
            "description": "an error has occurred",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/api/namespaces/{namespace}/tree/{workflow}?op=create-workflow": {
      "put": {
        "description": "Creates a workflow at the target path.\nThe body of this request should contain the workflow yaml.\n",
        "consumes": [
          "text/plain"
        ],
        "tags": [
          "Workflows"
        ],
        "summary": "Create a Workflow",
        "operationId": "createWorkflow",
        "parameters": [
          {
            "type": "string",
            "default": "create-workflow",
            "description": "the operation for the api",
            "name": "op",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "target namespace",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "path to target workflow",
            "name": "workflow",
            "in": "path",
            "required": true
          },
          {
            "description": "Payload that contains the direktiv workflow yaml to create.",
            "name": "workflow data",
            "in": "body",
            "schema": {
              "type": "string",
              "example": "description: A simple no-op state that returns Hello world!\nstates:\n- id: helloworld\n  type: noop\n  transform:\n    result: Hello world!\n"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfully created workflow",
            "schema": {
              "$ref": "#/definitions/OkBody"
            }
          },
          "default": {
            "description": "an error has occurred",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/api/namespaces/{namespace}/tree/{workflow}?op=delete-var": {
      "delete": {
        "description": "Delete a workflow variable.\n",
        "tags": [
          "Variables"
        ],
        "summary": "Delete a Workflow Variable",
        "operationId": "deleteWorkflowVariable",
        "parameters": [
          {
            "type": "string",
            "description": "target namespace",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "target variable",
            "name": "var",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "path to target workflow",
            "name": "workflow",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successfully deleted workflow variable"
          }
        }
      }
    },
    "/api/namespaces/{namespace}/tree/{workflow}?op=execute": {
      "post": {
        "description": "Executes a workflow with optionally some input provided in the request body as json.\n",
        "tags": [
          "Workflows"
        ],
        "summary": "Execute a Workflow",
        "operationId": "executeWorkflow",
        "parameters": [
          {
            "type": "string",
            "default": "execute",
            "description": "the operation for the api",
            "name": "op",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "target namespace",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "path to target workflow",
            "name": "workflow",
            "in": "path",
            "required": true
          },
          {
            "description": "The input of this workflow instance",
            "name": "Workflow Input",
            "in": "body",
            "schema": {
              "type": "object",
              "example": {
                "animals": [
                  "dog",
                  "cat",
                  "snake"
                ]
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "node has been deleted",
            "schema": {
              "$ref": "#/definitions/OkBody"
            }
          },
          "default": {
            "description": "an error has occurred",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/api/namespaces/{namespace}/tree/{workflow}?op=logs": {
      "get": {
        "description": "Get workflow level logs.\n",
        "tags": [
          "Logs"
        ],
        "summary": "Get Workflow Level Logs",
        "operationId": "getWorkflowLogs",
        "parameters": [
          {
            "type": "string",
            "description": "target namespace",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "path to target workflow",
            "name": "workflow",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successfully got workflow logs"
          }
        }
      }
    },
...
    "/api/namespaces/{namespace}/tree/{workflow}?op=set-var": {
      "put": {
        "description": "Set the value sorted in a workflow variable.\nIf the target variable does not exists, it will be created.\nVariable data can be anything so long as it can be represented as a string.\n",
        "consumes": [
          "text/plain"
        ],
        "tags": [
          "Variables"
        ],
        "summary": "Set a Workflow Variable",
        "operationId": "setWorkflowVariable",
        "parameters": [
          {
            "type": "string",
            "description": "target namespace",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "target variable",
            "name": "var",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "path to target workflow",
            "name": "workflow",
            "in": "path",
            "required": true
          },
          {
            "description": "Payload that contains variable data.",
            "name": "data",
            "in": "body",
            "schema": {
              "type": "string",
              "example": {
                "counter": 0
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfully set workflow variable"
          }
        }
      }
    },
    "/api/namespaces/{namespace}/tree/{workflow}?op=set-workflow-event-logging": {
      "post": {
        "description": "Set Cloud Event for Workflow to Log to.\nWhen configured type `direktiv.instanceLog` cloud events will be generated with the `logger` parameter set to the configured value.\nWorkflows can be configured to generate cloud events on their namespace anything the log parameter produces data.\nPlease find more information on this topic here:\nhttps://docs.direktiv.io/docs/examples/logging.html\n",
        "tags": [
          "Workflows"
        ],
        "summary": "Set Cloud Event for Workflow to Log to",
        "operationId": "setWorkflowCloudEventLogs",
        "parameters": [
          {
            "type": "string",
            "description": "target namespace",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "path to target workflow",
            "name": "workflow",
            "in": "path",
            "required": true
          },
          {
            "description": "Cloud event logger to target",
            "name": "Cloud Event Logger",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "logger"
              ],
              "properties": {
                "logger": {
                  "description": "Target Cloud Event",
                  "type": "string"
                }
              },
              "example": {
                "logger": "mylog"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfully update workflow"
          }
        }
      }
    },
    "/api/namespaces/{namespace}/tree/{workflow}?op=toggle": {
      "post": {
        "description": "Toggle's whether or not a workflow is active.\nDisabled workflows cannot be invoked. This includes start event and scheduled workflows.\n",
        "tags": [
          "Workflows"
        ],
        "summary": "Set Cloud Event for Workflow to Log to",
        "operationId": "toggleWorkflow",
        "parameters": [
          {
            "type": "string",
            "description": "target namespace",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "path to target workflow",
            "name": "workflow",
            "in": "path",
            "required": true
          },
          {
            "description": "Whether or not the workflow is alive or disabled",
            "name": "Workflow Live Status",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "live"
              ],
              "properties": {
                "live": {
                  "description": "Workflow live status",
                  "type": "boolean"
                }
              },
              "example": {
                "live": false
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfully updated workflow live status"
          }
        }
      }
    },
    "/api/namespaces/{namespace}/tree/{workflow}?op=update-workflow": {
      "post": {
        "description": "Updates a workflow at the target path.\nThe body of this request should contain the workflow yaml you want to update to.\n",
        "consumes": [
          "text/plain"
        ],
        "tags": [
          "Workflows"
        ],
        "summary": "Update a Workflow",
        "operationId": "updateWorkflow",
        "parameters": [
          {
            "type": "string",
            "description": "target namespace",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "path to target workflow",
            "name": "workflow",
            "in": "path",
            "required": true
          },
          {
            "description": "Payload that contains the updated direktiv workflow yaml.",
            "name": "workflow data",
            "in": "body",
            "schema": {
              "type": "string",
              "example": "description: A simple no-op state that returns Hello world Updated !!!\nstates:\n- id: helloworld\n  type: noop\n  transform:\n    result: Hello world Updated !!!\n"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfully updated workflow"
          }
        }
      }
    },
    "/api/namespaces/{namespace}/tree/{workflow}?op=var": {
      "get": {
        "description": "Get the value sorted in a workflow variable.\n",
        "tags": [
          "Variables"
        ],
        "summary": "Get a Workflow Variable",
        "operationId": "getWorkflowVariable",
        "parameters": [
          {
            "type": "string",
            "description": "target namespace",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "target variable",
            "name": "var",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "path to target workflow",
            "name": "workflow",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successfully got workflow variable"
          }
        }
      }
    },
    "/api/namespaces/{namespace}/tree/{workflow}?op=vars": {
      "get": {
        "description": "Gets a list of variables in a workflow.\n",
        "tags": [
          "Variables"
        ],
        "summary": "Get List of Workflow Variables",
        "operationId": "getWorkflowVariables",
        "parameters": [
          {
            "type": "string",
            "description": "target namespace",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "path to target workflow",
            "name": "workflow",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successfully got workflow variables"
          }
        }
      }
    },
...
    "/api/namespaces/{namespace}/vars": {
      "get": {
        "description": "Gets a list of variables in a namespace.\n",
        "tags": [
          "Variables"
        ],
        "summary": "Get Namespace Variable List",
        "operationId": "getNamespaceVariables",
        "parameters": [
          {
            "type": "string",
            "description": "target namespace",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successfully got namespace variables"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "Message": {
          "type": "string",
          "x-go-name": "Error"
        },
        "StatusCode": {
          "type": "integer",
          "format": "int64"
        }
      },
      "x-go-package": "github.com/vorteil/direktiv/pkg/api"
    },
...
    "OkBody": {
      "description": "OkBody OkBody ok body",
      "type": "object",
      "additionalProperties": {
        "type": "object"
      },
      "x-go-package": "github.com/vorteil/direktiv/pkg/api/models"
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "KEY",
      "in": "header"
    }
  },
  "security": [
    {
      "api_key": []
    }
  ]
}